cellquant-v3/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cellquant"
version = "3.0.0"
description = "High-throughput cell quantification and tracking for microscopy"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2",
    "numpy>=1.24",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "tifffile>=2024.1",
    "scikit-image>=0.22",
    "scipy>=1.12",
    "Pillow>=10.2",
]

[project.optional-dependencies]
gpu = ["cellpose>=3.0", "torch>=2.1"]
tracking = ["trackastra>=0.1"]
napari = ["napari>=0.4.19", "pyqt5>=5.15"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
]
all = ["cellquant[gpu,tracking,napari]"]

[project.scripts]
cellquant = "cellquant.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cellquant"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
